# swagger/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "这里是swagger中整个项目的描述",
        "termsOfService": "https://www.swagger.io/terms/",
        "contact": {
            "name": "维护者名字",
            "url": "http://www.swagger.io/support",
            "email": "维护者邮件"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "版本(默认1.0)"
    },
    "host": "127.0.0.1:8080",
    "basePath": "/api/v1",
    "paths": {
        "/login": {
            "post": {
                "description": "login",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "测试2"
                ],
                "summary": "测试login",
                "parameters": [
                    {
                        "description": "用户信息",
                        "name": "Object",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/main.RequestParam"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/main.ResponseParam"
                        }
                    },
                    "400": {
                        "description": "{\"msg\": \"failed\"}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/ping": {
            "get": {
                "description": "ping",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "测试1"
                ],
                "summary": "测试ping",
                "responses": {
                    "200": {
                        "description": "{\"msg\": \"pong\"}",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "{\"msg\": \"nonono\"}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/update": {
            "put": {
                "description": "update",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "测试4"
                ],
                "summary": "更新用户信息",
                "parameters": [
                    {
                        "description": "用户更新后信息",
                        "name": "Object",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/main.RequestParam2"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/main.ResponseParam"
                        }
                    },
                    "400": {
                        "description": "{\"msg\": \"failed\"}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/user": {
            "delete": {
                "description": "deleteuser",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "测试3"
                ],
                "summary": "删除用户",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "用户id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {}
            }
        }
    },
    "definitions": {
        "main.RequestParam": {
            "type": "object",
            "properties": {
                "password": {
                    "type": "string",
                    "example": "123456"
                },
                "username": {
                    "type": "string",
                    "example": "李四"
                }
            }
        },
        "main.RequestParam2": {
            "type": "object",
            "properties": {
                "password": {
                    "type": "string",
                    "example": "123456"
                },
                "updatepassword": {
                    "type": "string",
                    "example": "789100"
                },
                "updateusername": {
                    "type": "string",
                    "example": "张三"
                },
                "username": {
                    "type": "string",
                    "example": "李四"
                }
            }
        },
        "main.ResponseParam": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {
                    "$ref": "#/definitions/main.User"
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "main.User": {
            "type": "object",
            "properties": {
                "password": {
                    "type": "string",
                    "example": "123456"
                },
                "username": {
                    "type": "string",
                    "example": "张三"
                }
            }
        }
    }
}